chore: some dev utilies
I run Postgres in the $(pwd)/data directory and all the build artifacts will be
installed to $(pwd)/install. So add these 2 directories to .gitignore. A
justfile is also created for command shortcuts I currently rely on. we do not
use using the existing Makefile to avoid potential conflicts with the
upstream.